+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-- Let user add file types to ignore when parsing (#21).
+- Support include and exclude options (#21).
 
 ## [0.3.1] - 2024-01-19